Change int damage to float damage

+++ FrontlineWalker/Assets/Scripts/TestScripts/TestCharacterPart.cs
@@ -8,14 +8,14 @@ public class TestCharacterPart : MonoBehaviour
     [Header("Settings")]
     [SerializeField] private String name;
     [Range(0, 1000)]
-    [SerializeField] private int maxHealth = 10;
+    [SerializeField] private float maxHealth = 10;
     [SerializeField] private bool isCritical;
     public TestMessangeSender messanger;
     [Header("Debug")] 
     public bool alwaysSendMessage = true;
     public bool showDebugLog = true;
     
-    private int _health;
+    private float _health;
     public bool _isWorking;
 
     private void FixedUpdate()
@@ -63,7 +63,7 @@ public class TestCharacterPart : MonoBehaviour
         }
     }
      
-    public int GetHealth()
+    public float GetHealth()
     {
         return _health;
     }
@@ -79,7 +79,7 @@ public class TestCharacterPart : MonoBehaviour
         CheckHealth();
     }
     
-    public void TakeDamage(int _damageAmount)
+    public void TakeDamage(float _damageAmount)
     {
         if (_damageAmount < 0)
         {
